Move ModalInteraction into correct package (#2282)
